# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cloud-Ready School Management ERP API - Phase 1",
        "description": "Complete RESTful API specifications for Multi-Campus School Management ERP. Encompasses Institute Management, Multi-Campus Branches, Academic Years & Terms, Spatie Role-Based Access Control (RBAC), Global Settings, Classes, Sections, Streams, Houses, Subjects, Employee & Staff Management, and Student Admissions & Profiles.",
        "version": "1.0.0",
        "contact": {
            "name": "School ERP Support",
            "email": "support@schoolerp.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Laravel Development API Server"
        }
    ],
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Sanctum API Bearer Token"
            }
        }
    },
    "paths": {
...
        "/staff": {
            "get": {
                "tags": [
                    "Staff & Employee Records"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "List Staff / Employees",
                "responses": {
                    "200": {
                        "description": "Paginated staff"
                    }
                }
            },
            "post": {
                "tags": [
                    "Staff & Employee Records"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Create Staff Record with Portal Login",
                "responses": {
                    "201": {
                        "description": "Staff created"
                    }
                }
            }
        },
        "/staff/{id}": {
            "get": {
                "tags": [
                    "Staff & Employee Records"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Get Staff Profile Details",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Staff profile"
                    }
                }
            },
            "put": {
                "tags": [
                    "Staff & Employee Records"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Update Staff Record",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Staff updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Staff & Employee Records"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Delete Staff Record",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Staff deleted"
                    }
                }
            }
        },
        "/staff/{staffId}/assign-subject": {
            "post": {
                "tags": [
                    "Staff & Employee Records"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Assign Teacher to Class, Section and Subject",
                "parameters": [
                    {
                        "name": "staffId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Teacher assignment recorded"
                    }
                }
            }
        },
        "/departments": {
            "get": {
                "tags": [
                    "Staff & Employee Records"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "List Departments",
                "responses": {
                    "200": {
                        "description": "List of departments"
                    }
                }
            },
            "post": {
                "tags": [
                    "Staff & Employee Records"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Create Department",
                "responses": {
                    "201": {
                        "description": "Department created"
                    }
                }
            }
        },
        "/designations": {
            "get": {
                "tags": [
                    "Staff & Employee Records"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "List Designations",
                "responses": {
                    "200": {
                        "description": "List of designations"
                    }
                }
            },
            "post": {
                "tags": [
                    "Staff & Employee Records"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Create Designation",
                "responses": {
                    "201": {
                        "description": "Designation created"
                    }
                }
            }
        },
...
    }
}